ML/AWS_production/requirements.txt
requests
boto3
numpy
Pillow
torchvision
torch